libcblite-*: update `livecheck` and `uninstall_postflight`

+++ Casks/l/libcblite.rb
@@ -8,8 +8,8 @@ cask "libcblite" do
   homepage "https://www.couchbase.com/products/lite"
 
   livecheck do
-    url "http://appcast.couchbase.com/couchbase-lite-version.txt"
-    regex(/(\d+(?:\.\d+)+)/i)
+    url "https://docs.couchbase.com/couchbase-lite/current/c/gs-install.html"
+    regex(/href=.*?couchbase[._-]lite[._-]c[._-]enterprise[._-]v?(\d+(?:\.\d+)+)[._-]macos\.zip/i)
   end
 
   conflicts_with cask: "libcblite-community"
@@ -27,8 +27,10 @@ cask "libcblite" do
   end
 
   uninstall_postflight do
-    puts "Removing library symlinks in #{HOMEBREW_PREFIX}/lib"
-    File.unlink("#{HOMEBREW_PREFIX}/lib/libcblite.#{version.major}.dylib", "#{HOMEBREW_PREFIX}/lib/libcblite.dylib")
+    if File.symlink?("#{HOMEBREW_PREFIX}/lib/libcblite.#{version.major}.dylib")
+      puts "Removing library symlinks in #{HOMEBREW_PREFIX}/lib"
+      File.unlink("#{HOMEBREW_PREFIX}/lib/libcblite.#{version.major}.dylib", "#{HOMEBREW_PREFIX}/lib/libcblite.dylib")
+    end
   end
 
   # No zap stanza required
